# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Credit Hub KE API - airtime & data sales with M-Pesa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update my profile (name)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, phone, balance, referralCode, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
          nullable: true
        balance:
          type: number
        referralCode:
          type: string
        createdAt:
          type: string
...
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
...
    Transaction:
      type: object
      required: [id, bundleId, bundleName, bundleType, recipientPhone, amount, status, createdAt]
      properties:
        id:
          type: integer
        bundleId:
          type: integer
        bundleName:
          type: string
        bundleType:
          type: string
        recipientPhone:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, failed]
        mpesaCode:
          type: string
          nullable: true
        createdAt:
          type: string
...